Move microfrontends to common folder
