+++ frontend/microfrontend/auth/src/bootstrap.js
