+++ frontend/microfrontend/auth/src/index.js
