+++ frontend/microfrontend/profile/src/components/EditAvatarPopup.js
