+++ frontend/microfrontend/profile/webpack.config.js
