--- frontend/profile-microfrontend/src/index.css
@@ -1,1 +1,0 @@
-@import url('./blocks/auth-form/auth-form.css');
